Add logic to send screening invites and reminders
Sending screening invitations and reminders requires not just sending mails but also storing data about the sent mails in the database. It is also used by both, the jobs and web part of the backend. Thus it is some kind of business logic which sits near the model and is not part of the controller (at least I don’t like that to put something like that into a controller).

The functions introduced with that commit are used in the whole backend now when someone tries to send screening invitations or reminders.