Update dev database setup
This commit will add the dev database setup to set the lastSentScreeningInvitationDate. This is set to the verifiedAt date, because screening invitations are sent after the email address was verified. And because both test students have a verification of null at the moment, they have verified their email address and thus should have received a screening invitation, at least one.