Fix bug where instructors got screening reminder
The usual screening reminders for 1-on-1 tutoring should be sent only to those users ("students“) who actually have the 1-on-1-tutoring role. Instructors shouldn’t get the same reminders as tutors. This is fixed with that commit. Now only students/tutors registered for 1-on-1 tutoring will receive the screening reminders.