Store tutor screening result for instructor separate
This commit will add the functionality required to appropriately store the „qualified for 1-on-1 tutoring“ information that is collected as part of the usual instructor screening process. When someone screens an instructor that isn’t already screened as tutor, it will now create an additional row in the 1-on-1-tutor-screening table. That means if an instructor is screened, two screening results will be stored in the database: one storing the screening result for being an instructor, the other for being a tutor.