Explicitly set moment timezone for jobs
Specifying the timezone „Europe/Berlin“ explicitly using moment-timezone in the jobs part of our backend to not to time related operations in the timezone of the server but instead in Germany’s timezone (because our users are in Germany).
This should fix that wrong dates (not in Germany’s timezone) were sent in course reminders.
This should also fix #93, because the cron package also internally seems to use moment-timezone.