Fix lecture started/ended computation
Fix the computation whether a lecture has started/ended or not to conform to the new real unix timestamp the API doc specifies for a ApiLecture. The old computation no longer works, because the previous commit has introduced the real unix timestamp (i.e. seconds, not milliseconds). So we needed to handle this, which is done with this commit.