Add screening date to verify-certificate page
As Tobi has noted: If we state „Namen und Studierendenstatus haben wir überprüft“ in the current version of the certifcate verification page, this could be misleading if someone accesses the webpage in 10 (just an example) years because it looks like that the person is still studying even though she might not.
The screening date is introduced such that we can take care of that by bringing in a temporal component.