Fix ParticipationCertificate date bug
In a previous version there existed the following bug: the ParticipationCertificate’s „certificateDate“, „startDate“, and „endDate“ fields were always set to a default value of „now“ even though if programmatically set in code to the actual value that was selected as part of the certificate generation process. The problem was: Because of the usage of CreateDateColumn, these three columns were marked as special (with that special CreateDateColumn decorator). But columns with CreateDateColumn will always store the value of the then current datetime in the variable, independent of what programmatically was set. This is fixed with this commit. It will introduce what was probably originally intended: default now() and nullable false. This does not require a migration, because the database schema was also set correct. Moreover the problem was on application level in typeorm ➔ it always discarded values for any of that three columns and inserted the default vallues.

By the way: some other small bugs related to date were fixed with this commit, i.e. wrong parsing and assigning the end date to both start and end.