Add migration for course attendance log
This will add the typeorm database migration required to add the necessary entities (and changes on existing entities) for the course attendence log as well as modifying the enum that will store possible log types. Those two things are all database changes introduced as part of adding the course attendance log feature.