Update VSCode debug setup for in-previw debugger
This debug-setup (launch.json with corresponding package.json scripts) will work with the new in-preview Node debugger shipped with VSCode. Part of this is that the debug-scripts in package.json now use the „port“ 0, i.e. auto select the debugging port. This also changes the timeouts, required to compile all files with tsc.
It will also add a debug configuration for the new node-terminal in VSCode that uses tsnode to execute the web part.