Fix course join if not all join during 15 minutes
This commit will fix a problem that could occur for instructors and participants.
The problem: If an instructors starts the course and not all people join during 15 minutes (the interval after which the meeting cache in the backend is updated), participants will no longer be able to join. And also instructors will no longer be able to join again after the 15 minutes.

This problem must have been existent since we’re using the subcourse ID to join the meetings.

Steps to reproduce (before this fix):
An instructor starts a course, i.e. the BBB meeting. Wait the specified interval until BBB meeting cache updates in backend. Try to use the same „join course“ button in the frontend of the instructor again. This time it will fail, because the meeting cache’s values are not the same as those used to initialize the meeting. Participants have the same problem.

Reason:
This was caused by using the given subcourseID of the http-request (which was given as a number) as a parameter to the BBB methods and as a key for the BBB-Meeting-Cache. Even though the cache was typed as string, this doesn’t matter, because after compilations, it’s javascript that’s executed… So the meeting cache got invalid after 15 minutes, overwritten with the current info from the BBB server, now correctly interpreted as a string. So a running meeting was then never again found.