Quick fix non editable subjects for tutees
This commit should fix corona-school/web-user-app#81, which initially thought to be a frontend problem.

This will add a method that strips off grade information from tutee’s subject data fetched from the database that contains grade information (even if it shouldn’t have grade information cause they’re tutees). This additional grade information (if returned to the frontend) lead to issues when a user is editing his subject data in the frontend. Because then the frontend will make a PUT with an array of subject where some subjects have grade information and some have not. Because the backend expects all subjects in the array have the same format, it will throw an error in these cases. So we need to consistently return tutee’s subjects only without grade information.
Note: Tutee’s subject were written into the database with grade information (i.e. WRONG) during registration (which will then cause the problem described above). So this should be fixed in a separate commit in the future.