Add registration for project coaching
This commit will introduce the basic extension of the existing registration API endpoints. It allows registration for project coaches (students/tutors) and for project mentees (pupils/tutees).

Part of this is the extension of the data model. Note: the database migration is still missing.

The following smaller things are also part of this commit:
- the „reverse mapping“ for string enums is now more generalized to be simpler to reuse and have a cleaner code