Change way how project fields are stored for tutors
We need to change the way how project fields are stored for tutors because the system needs to maintain information on which grade restrictions are set for which project fields of a tutor. These grade restrictions should be set during the screening. Therefore this commit introduces changes which use a separate entity/table to store project fields including grade restrictions for a tutor.