Extend coach registration to store jufo cert info
The backend API for 1-on-1-project-coaches now expects info on whether the participant has a certificate that verifies participation in past jufo-competitions iff the coach is no university student, but a past jufo participant. This information on whether the coach has a certificate or not is also stored in the database.