Update screening email handling (inlcuding jufo)
This commit does several things related to the screening reminders with the focus on implementing reminders for the screening interviews that are required as part of the project coaching/jufo stuff.
First, this will update the way how remindees are calculated such that this is more general and intuitive to better fit the needs for the project coaching screening reminders.
As part of this, it was recognized that the reminders for instructors weren’t implemented at all, at least not implemented as expected. This commit will thus also add screening reminders for instructors.
It also implements initial/first screening invitation mails for project coaching screening interviews.