Refactor screening to a single endpoint
This commit will introduce various changes that are related to making the screening API (that is used by the screening tool to store infos on the tutors) simpler, easier to use and more reliable.
It should make implementing the frontend simpler. The frontend should now do all the screenings in the single queue while at the same time introducing Jufo-screenings. So this commit should help supporting that. It will also allow editing various other settings of a student, which is also intended to become a new feature in the screening tool.