Change openMatchRequestCount value if not tutor
When someone registers not as a tutor, until now that person would have got an openMatchRequestCount of zero (because she’s not a tutor). This commit will always use the default value for openMatchRequestCount, i.e. a value of 1, independent of whether someone is a tutor or only registering for project coaching. This will allow the following even simpler: if someone registers as a project coach (or as an instructor) and not as a tutor, but a screener can convince them to be a 1-on-1 tutor, we can set the isStudent variable of that person to true and then he will automatically get a first match. This is desired, because we need way much more 1-on-1 learning tutors.