Add log statement to screener API’s main method
This log statement will log the info that someone tries to update a student with. We’re logging this info to have it in case of something goes wrong. In the past it happened that the screening tool was kinda buggy and which sent buggy requests to the API, but everyone assumed that the screening was sucessfully done. In those cases we have missed the main info (i.e. if someone is verified), even in those cases where only the grade restrictions were malformed. So if we have a logging statement, we can go back to our logs and see the buggy requests including the most important information that should have been saved.