Fix impossible matching if no grade was provided
For project coachees it's possible to provide no grade during registration, i.e. for example required for "Berufsschule" or students participating in Jugend forscht. To still match them with our usual matching process, we need to assume a default grade. This commit will add a default grade of 13 for those cases, because 13 is the oldest grade and the nearest grade to university students.