Add new endpoint for adding instructor to course
This will implement a very basic and separate endpoint that can be used to add another person as an instructor to a course. That other person must fulfill certain conditions. This endpoint can be called by another instructor who already has access rights to the appropriate course.
It basically implements the backend for corona-school/project-user#228.