Change S3 endpoint handling to make tests work
Test won't work if no endpoint information is provided. This commit provides a quick fix for this problem by using the default AWS S3 endpoint if no one was provided manually. There may be better solutions and the entire aws-sdk handling and how it is tested and used during development on your local machine needs to be revised in the future.