Add course level allow contact option
This basically is the backend part for corona-school/project-user#196 and adds a new property to the course entity which stores whether participants are allowed to contact all instructor of a course (i.e. "on a course level").
The API endpoints for posting and putting courses are also adjusted. Also the endpoints for retrieving courses now deliver that information on whether contacting course instructors is allowed.

The allowContact property might be adjusted to use an array of instructor IDs instead of a single boolean value, depending on https://github.com/corona-school/project-user/issues/195#issuecomment-755855440 and https://github.com/corona-school/project-user/issues/196#issuecomment-755854579.