Improve instructor search by firstname
As far as I can see, the instructor review page for which this search was originally built was replaced with something else.
This API endpoint is now only used for the "Manage Instructors" dialogue, as far as I can see.
In that dialogue, it seems as if one can search instructors by full name, however that is not quite true.
So I thought to finish up the "search in review" topic, I'll align the behavior of this search with the one in the courses search, allowing users to search by fullname.