Make sender and subject option in mail templates
This way, it will be possible to have a template defined here in the backend code which does not define sender and subject/title. Then mailjet will use the sender and title defined in it's template editing frontend. So it's easier to maintain templates, while the old option still continues to work.