Add waiting list to subcourses
This commit will add a waiting list to all subcourses. Participants can use the corresponding two new API endpoints introduced with this commit to join/leave a waiting list for a subcourse.
If a pupil tries to join a course where he's already on the waiting list, the join will be completed and he will be removed from the waiting list.
Corresponding issue for reference: corona-school/project-user#220