Add basic test endpoint for test BBB meetings
This commit adds a new API endpoint which is publicly readable and on every request returns a redirection to a new empty BBB meeting. Someone can then participate in this meeting and test whether his connection to our (production) server works. That person should click on the button to join by microphone and check whether the echo test in BBB can be completed with success.