Implement backend part for correspondent contact
How participants can contact instructors will be changed with this commit. Now course instructors can define a so called "correspondent" for a course who will receive all contact requests of pupils. This is still only possible if allowContact is switched on for that specific course. See corona-school/project-user#195 and corona-school/project-user#196.