Extend bbb test endpoint with user authentication
The API endpoint that creates a test meeting for testing the BBB-Server will now also support authentication using the usual Token Header. If a user (either student or pupil) will use the endpoint with a token, he will receive a test meeting that is unique for him, but that will be the same across multiple calls to the endpoint. If that endpoint is called without an authentication token, it will, as previously, still return a new meeting where someone can join with a random name.