Implement invitation of external guests to courses
This will be the backend part for corona-school/project-user#241. It will add two new endpoints. One will allow course instructors to invite at max 5 guests to a course and send an email to the invited people that will contain a link they can use to join the course. The second endpoint will be publicly available and can be used with links sent to various people as part of the invitation process, it will then return the BBB url that can be used to join a course meeting.