Improve screening api's getCourseTags performance
This will prevent some massive eager loading (i.e. all courses with all corresponding users etc.) when trying to call the getTags endpoint for the screeners. This massive eager loading was not used and lead to huge performance problems, leading to veryyyy long queries and request as well as to very frequent occurrence of out-of-memory errors. This commit will fix this by using the query API that prisma provides. But note: the prisma code is still neither very good tested nor very beautiful. For example one problem is the ESLint stuff…