Modify build:clean command to support devcontainer
If devcontainers are used with a mounted built directory, that directory cannot be removed like all the other non mounted directories. Therefore if we wanna have a clean build, we need to delete all the contents of this folder instead. Therefore this commit adds a fallback mechanism in case the `rm -rf built/` command fails (the error output of this command is discarded because of the fallback mechanism).
This commit also changes the build:clean command to refer to the actual build command instead of calling `tsc` on its own.