Disable eager loading courses during student auth
When a student tried to use an auth-token to authenticate against our backend API until now all courses of this student, including all eager properties in the chain, loaded automatically on EVERY authentication. This is quite useless, because it is no longer used and thus it only takes time, memory and computation power. Bringing this together with some situations where typeorm has memory problems (which usually occur due to a lot of nested eager loads in entities related to our courses), it is a dramatic improvement to remove this automatic "eager" course loading, which this commit does.