Disable match follow up and feedback request jobs
Those two jobs were implemented in the past, but actually not really used in production for a couple of operational reasons originating from the content perspective. Thus we're going to disable them even though the basic structure remains intact. Maybe we will reactivate them in the future. But from now on, those jobs are no longer scheduled, thus the corresponding column values on the Match entity are no longer important in any way.