Add quite generic manual tutoring matching endpoint
This commit will extend the backend api by an additional new endpoint which can be used to manually trigger matching on the backend. This is part of the screening API routes in a way that it is only accessible with a valid Screener-auth-token. This commit will now make the external match creation tools part of the backend code. It still won't mean automatic matching, but the new endpoint should provide a simpler interface to fulfill a lot of the common manual matching requests.

More things to note in relation to this commit:
* This endpoint allows only to make valid matches where "valid" means for example that it will not match unscreened tutors.
* It will also use 'class-transformer' and 'class-validator' packages to define and validate the API-requests.
* Concurrency management when working with the database needs improvements throughout the code, so this endpoint cannot guarantee that only valid matches were made, notified and saved if some concurrency race conditions occur.