Fix bug on joining BBB meeting
When a course name includes an "'" (apostrophe/single quote) character, the corresponding BBB meeting name also includes this. Previously this single quote was encoded in a way that differs from what BBB expects. So this commit will fix this encoding issue such that now single quotes are supported in course names too. This wasn’t the case previously where the meeting creation request was sent to BBB but wasn’t working appropriately.

Also error handling is improved with this commit. If an API request for joining a meeting is sent to the BBB server, the error handling previously was not really existent. This made it possible for the user to have been redirected to BBB even though the actual meeting creation failed. So if a meeting name contained a single quote (as above), the user was redirected to the wrong BBB page due to this missing error management instead of indicating an internal error.

But the whole bbb stuff needs some refactoring…