Fix limit bug when using QueryBuilder from typeorm
Also see https://github.com/typeorm/typeorm/issues/4883.
This is the same as the undefined check here, which this commit also adds, but which is not sufficient for solving the problem of a limit that does not apply if set to zero.
So this commit introduces a "workaround" (or shortcut…) around typeorm’s bug. Previously pupils were invited even though the limit computation said that no (i.e. 0) pupils should get the interest confirmation request.