feature: move GraphQL error logging to formatError
Unfortunately, the didEncounterErrors hook doesn't have a lot of information about which kind of errors did happen.
This means, that it's alsomost impossible to distinguish them and decide which to log how.

Therfore, I moved the logging to the formatError hook, because the given error is always wrapped into a class with context, like ValidationError.
This helps us to devide when to log what.