Make User non-optional and readonly
We set it to some fake user anyways in most cases, so we can also do it in all cases.

We currently write the roles into the user and potentially modify it, refactored to then always clone the whole user.

isSessionStudent et. al. type guard the GraphQLContext, to get rid of lots of non null assertions.