fix: Retry Lock if rolled back by DB
If the DB transaction succeeded we know if another job is running or not and can abort / continue accordingly.

If the transaction was rolled back by the DB due to a deadlock, we do not know if two transactions on the same job name collided or unrelated transactions collided (due to limitations in Postgres deadlock detection). Thus in that case it makes sense to retry, as usually one concurrent transaction should be aborted, so if we try again there should be no concurrent transaction