feat: Include message in push notification
It is not so easy to establish a backend connection in the webworker - thus it seems easier to just include the whole message to show in the push notification itself. As we decrypt the message using the client key, only the client can read it - so there should be no privacy problems - as users need to trust their browsers anyways.