feat: Add NotAllowedError class for handling forbidden actions
The code changes introduce a new class called NotAllowedError, which extends the Error class. This class is used to handle forbidden actions and send an appropriate error message to the client. The constructor of the NotAllowedError class takes two parameters: type and publicMessage. The type parameter represents the type of the error, while the publicMessage parameter contains the error message that will be sent to the client. The commit also includes some minor formatting changes.

Co-authored-by: Jonas Wilms <jonas@wilms.ninja>