feat: Add retention
Keep pupils a bit longer in the match pool to find a better match. This increases wait time from 4 days to 17 days on average, but increases both the number of subjects and states matched. Not sure if the trade-off is worth it, but now we have a mechanism for it that can be tuned.