fix: Reduce likelihood of race condition in Matching
In case two users are matched at the same time (e.g. through a Match run + manual matching) there is a certain likelihood that they are matched twice.
By refetching the openMatchRequestCount right before creating the match, the likelihood of that happening is reduced.

There might also be a bug in the new matching that creates duplicate matches, I hope if there is one, we will find it with this change.