fix: SSO_REGISTERING_USER is OWNER
Previously SSO_REGISTERING_USER (in case of link / register) would not include the USER role, however it would imply the OWNER role on the user entity. This means that the many resolvers guarded by OWNER could be run by such a temporary user. I am not certain that this correctly fails in all cases (in particular this runs into getUser(user.userID), which fails for a mock user-ID the sso user has), thus I entirely removed the OWNER role from such users by default, adding the TEMPORARY_OWNER role to allow the few cases where we want to allow this, i.e. the parts that are in the GetUser query the User-App uses globally

https://github.com/corona-school/project-user/issues/1488