fix: Various (#1303)
* fix: No actionUrls if user is no longer invitee of Calendly Event

Looks like the user registered with an invalid email, which was fixed by Support - thus the email mismatched.

* fix: Make it clear that error could also be Admin

There was a wrong query on the Retool Kurse causing silent errors in the background.

* fix: No firstname/lastname during Google SSO

Users can remove their firstname/lastname in their Google Profile, during SSO this is null then - which causes chaos in our session information, as we require firstname/lastname to not be null. Thus set them to an empty string instead. Users can later set that in their profile / during registration then.

* followup

* linter hapiness

---------

Co-authored-by: Jonas Wilms <jonas@wilms.ninja>